Week3 html project
-A modest proposal
I searched lots of references from internet to do make better. I added
some lines and photos. Also figured out how to change the type face
into italics and change color. The part that I struggled the most is
that I really wanted to make the paragraph in the center and I finally
figured it out! Although it was quite a challenge but I’m so glad that
I finally figured it out

-Typespecimen
The text I used to see the different size was the lyrics of “Sunday
Morning” from Maroon 5. I put different font size in different
paragraph and change the type face throughout the column. I used three
different type face. It was very hard to set three different font to
each different column.